requirements.txt
aiogram>=3.6
python-dotenv>=1.0
openai>=1.40
pillow>=10.3
pytesseract>=0.3.10
rapidfuzz>=3.9
